pyproject.toml
[tool.poetry]
name = "e-commerce-platform"
version = "0.1.0"
description = ""
authors = ["RocketMenace <grizzlygit777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
django-environ = "^0.12.0"
django-countries = "^7.6.1"
pytest-django = "^4.10.0"
ipython = "^8.32.0"
djangorestframework-simplejwt = "^5.4.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
